Plan 9 from Bell Labs 2003-04-19

+++ sys/src/ape/lib/ap/stdio/snprintf.c
@@ -2,7 +2,7 @@
  * pANS stdio -- sprintf
  */
 #include "iolib.h"
-int snprintf(char *buf, int nbuf, const char *fmt, ...){
+int snprintf(char *buf, size_t nbuf, const char *fmt, ...){
 	int n;
 	va_list args;
 	FILE *f=_IO_sopenw();
